generate captcha

+++ magickwand/image.h
@@ -3,7 +3,7 @@
 #include <wand/MagickWand.h>
 
 void draw_on_buf(
-  void* buf, size_t x, size_t y, size_t width, size_t height, double angle,
+  void* buf, int x, int y, size_t width, size_t height, double angle,
   size_t font_size,
   const char* fgcolor,
   const char* font,
